Add AbstractResponse, refactor concrete *Response classes.  Resolves #344.
