+++ src/main/java/org/imsglobal/caliper/entities/response/AbstractResponse.java
@@ -1,0 +1,136 @@
+package org.imsglobal.caliper.entities.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.imsglobal.caliper.entities.AbstractEntity;
+import org.imsglobal.caliper.entities.EntityType;
+import org.imsglobal.caliper.entities.TimePeriod;
+import org.imsglobal.caliper.entities.resource.Attempt;
+import org.imsglobal.caliper.validators.EntityValidator;
+import org.joda.time.DateTime;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+/**
+ * This class provides a skeletal implementation of the Response interface
+ * in order to minimize the effort required to implement the interface.
+ */
+public class AbstractResponse extends AbstractEntity implements CaliperResponse {
+
+    @JsonProperty("attempt")
+    private Attempt attempt;
+
+    @JsonIgnore
+    private TimePeriod timePeriod = new TimePeriod();
+
+    /**
+     * @param builder apply builder object properties to the Response object.
+     */
+    protected AbstractResponse(Builder<?> builder) {
+        super(builder);
+
+        EntityValidator.checkStartTime(builder.timePeriod.getStartedAtTime(), builder.timePeriod.getEndedAtTime());
+        EntityValidator.checkDuration(builder.timePeriod.getDuration());
+
+        this.attempt = builder.attempt;
+        this.timePeriod.setStartedAtTime(builder.timePeriod.getStartedAtTime());
+        this.timePeriod.setEndedAtTime(builder.timePeriod.getEndedAtTime());
+        this.timePeriod.setDuration(builder.timePeriod.getDuration());
+    }
+
+    /**
+     * @return attempt associated with the response;
+     */
+    @Nonnull
+    public Attempt getAttempt() {
+        return attempt;
+    }
+
+    /**
+     * @return started at time
+     */
+    @Nullable
+    public DateTime getStartedAtTime() {
+        return timePeriod.getStartedAtTime();
+    }
+
+    /**
+     * @return ended at time
+     */
+    @Nullable
+    public DateTime getEndedAtTime() {
+        return timePeriod.getEndedAtTime();
+    }
+
+    /**
+     * @return duration
+     */
+    @Nullable
+    public String getDuration() {
+        return timePeriod.getDuration();
+    }
+
+    /**
+     * Builder class provides a fluid interface for setting object properties.
+     * @param <T> builder
+     */
+    public static abstract class Builder<T extends Response.Builder<T>> extends AbstractEntity.Builder<T>  {
+        private Attempt attempt;
+        private TimePeriod timePeriod = new TimePeriod();
+
+        /**
+         * Initialize type with default value.
+         */
+        public Builder() {
+            super.type(EntityType.RESPONSE);
+        }
+
+        /**
+         * @param attempt
+         * @return builder.
+         */
+        public T attempt(Attempt attempt) {
+            this.attempt = attempt;
+            return self();
+        }
+
+        /**
+         * @param startedAtTime
+         * @return
+         */
+        public T startedAtTime(DateTime startedAtTime) {
+            this.timePeriod.setStartedAtTime(startedAtTime);
+
+            return self();
+        }
+
+        /**
+         * @param endedAtTime
+         * @return builder
+         */
+        public T endedAtTime(DateTime endedAtTime) {
+            this.timePeriod.setEndedAtTime(endedAtTime);
+            return self();
+        }
+
+        /**
+         * @param duration
+         * @return
+         */
+        public T duration(String duration) {
+            this.timePeriod.setDuration(duration);
+            return self();
+        }
+    }
+
+    /**
+     *
+     */
+    private static class Builder2 extends Response.Builder<Builder2> {
+        @Override
+        protected Builder2 self() {
+            return this;
+        }
+    }
+}
