+++ src/main/java/org/imsglobal/caliper/entities/response/MultipleResponseResponse.java
@@ -37,39 +37,17 @@ import java.util.List;
  * Representation of a response to a multiple choice question that permits one or more
  * options to be selected.
  */
-public class MultipleResponseResponse extends AbstractEntity implements CaliperResponse {
-
-    @JsonProperty("attempt")
-    private Attempt attempt;
+public class MultipleResponseResponse extends Response {
 
     @JsonProperty("values")
     private ImmutableList<String> values;
-
-    @JsonIgnore
-    private TimePeriod timePeriod = new TimePeriod();
 
     /**
      * @param builder apply builder object properties to the Response object.
      */
     protected MultipleResponseResponse(Builder<?> builder) {
         super(builder);
-
-        EntityValidator.checkStartTime(builder.timePeriod.getStartedAtTime(), builder.timePeriod.getEndedAtTime());
-        EntityValidator.checkDuration(builder.timePeriod.getDuration());
-
-        this.attempt = builder.attempt;
         this.values = ImmutableList.copyOf(builder.values);
-        this.timePeriod.setStartedAtTime(builder.timePeriod.getStartedAtTime());
-        this.timePeriod.setEndedAtTime(builder.timePeriod.getEndedAtTime());
-        this.timePeriod.setDuration(builder.timePeriod.getDuration());
-    }
-
-    /**
-     * @return attempt associated with the response;
-     */
-    @Nonnull
-    public Attempt getAttempt() {
-        return attempt;
     }
 
     /**
@@ -81,35 +59,10 @@ public class MultipleResponseResponse extends AbstractEntity implements CaliperR
     }
 
     /**
-     * @return started at time
-     */
-    @Nullable
-    public DateTime getStartedAtTime() {
-        return timePeriod.getStartedAtTime();
-    }
-
-    /**
-     * @return ended at time
-     */
-    @Nullable
-    public DateTime getEndedAtTime() {
-        return timePeriod.getEndedAtTime();
-    }
-
-    /**
-     * @return duration
-     */
-    @Nullable
-    public String getDuration() {
-        return timePeriod.getDuration();
-    }
-
-    /**
      * Builder class provides a fluid interface for setting object properties.
      * @param <T> builder
      */
-    public static abstract class Builder<T extends Builder<T>> extends AbstractEntity.Builder<T>  {
-        private Attempt attempt;
+    public static abstract class Builder<T extends Builder<T>> extends Response.Builder<T>  {
         private List<String> values = Lists.newArrayList();
         private TimePeriod timePeriod = new TimePeriod();
 
@@ -118,15 +71,6 @@ public class MultipleResponseResponse extends AbstractEntity implements CaliperR
          */
         public Builder() {
             super.type(EntityType.MULTIPLERESPONSE);
-        }
-
-        /**
-         * @param attempt
-         * @return builder.
-         */
-        public T attempt(Attempt attempt) {
-            this.attempt = attempt;
-            return self();
         }
 
         /**
@@ -144,34 +88,6 @@ public class MultipleResponseResponse extends AbstractEntity implements CaliperR
          */
         public T value(String value) {
             this.values.add(value);
-            return self();
-        }
-
-        /**
-         * @param startedAtTime
-         * @return
-         */
-        public T startedAtTime(DateTime startedAtTime) {
-            this.timePeriod.setStartedAtTime(startedAtTime);
-
-            return self();
-        }
-
-        /**
-         * @param endedAtTime
-         * @return builder
-         */
-        public T endedAtTime(DateTime endedAtTime) {
-            this.timePeriod.setEndedAtTime(endedAtTime);
-            return self();
-        }
-
-        /**
-         * @param duration
-         * @return
-         */
-        public T duration(String duration) {
-            this.timePeriod.setDuration(duration);
             return self();
         }
