+++ src/main/java/org/imsglobal/caliper/entities/response/Response.java
@@ -18,122 +18,26 @@
 
 package org.imsglobal.caliper.entities.response;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.imsglobal.caliper.entities.AbstractEntity;
 import org.imsglobal.caliper.entities.EntityType;
-import org.imsglobal.caliper.entities.TimePeriod;
-import org.imsglobal.caliper.entities.resource.Attempt;
-import org.imsglobal.caliper.validators.EntityValidator;
-import org.joda.time.DateTime;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
-public class Response extends AbstractEntity implements CaliperResponse {
-    @JsonProperty("attempt")
-    private Attempt attempt;
-
-    @JsonIgnore
-    private TimePeriod timePeriod = new TimePeriod();
+public class Response extends AbstractResponse {
 
     /**
      * @param builder apply builder object properties to the Response object.
      */
-    protected Response(Builder<?> builder) {
-        super(builder);
-
-        EntityValidator.checkStartTime(builder.timePeriod.getStartedAtTime(), builder.timePeriod.getEndedAtTime());
-        EntityValidator.checkDuration(builder.timePeriod.getDuration());
-
-        this.attempt = builder.attempt;
-        this.timePeriod.setStartedAtTime(builder.timePeriod.getStartedAtTime());
-        this.timePeriod.setEndedAtTime(builder.timePeriod.getEndedAtTime());
-        this.timePeriod.setDuration(builder.timePeriod.getDuration());
-    }
-
-    /**
-     * @return attempt associated with the response;
-     */
-    @Nonnull
-    public Attempt getAttempt() {
-        return attempt;
-    }
-
-    /**
-     * @return started at time
-     */
-    @Nullable
-    public DateTime getStartedAtTime() {
-        return timePeriod.getStartedAtTime();
-    }
-
-    /**
-     * @return ended at time
-     */
-    @Nullable
-    public DateTime getEndedAtTime() {
-        return timePeriod.getEndedAtTime();
-    }
-
-    /**
-     * @return duration
-     */
-    @Nullable
-    public String getDuration() {
-        return timePeriod.getDuration();
-    }
+    protected Response(Builder<?> builder) { super(builder); }
 
     /**
      * Builder class provides a fluid interface for setting object properties.
      * @param <T> builder
      */
-    public static abstract class Builder<T extends Builder<T>> extends AbstractEntity.Builder<T>  {
-        private Attempt attempt;
-        private TimePeriod timePeriod = new TimePeriod();
+    public static abstract class Builder<T extends Builder<T>> extends AbstractResponse.Builder<T>  {
 
         /**
          * Initialize type with default value.
          */
         public Builder() {
-            super.type(EntityType.TRUEFALSE);
-        }
-
-        /**
-         * @param attempt
-         * @return builder.
-         */
-        public T attempt(Attempt attempt) {
-            this.attempt = attempt;
-            return self();
-        }
-
-        /**
-         * @param startedAtTime
-         * @return
-         */
-        public T startedAtTime(DateTime startedAtTime) {
-            this.timePeriod.setStartedAtTime(startedAtTime);
-
-            return self();
-        }
-
-        /**
-         * @param endedAtTime
-         * @return builder
-         */
-        public T endedAtTime(DateTime endedAtTime) {
-            this.timePeriod.setEndedAtTime(endedAtTime);
-            return self();
-        }
-
-        /**
-         * @param duration
-         * @return
-         */
-        public T duration(String duration) {
-            this.timePeriod.setDuration(duration);
-            return self();
+            super.type(EntityType.RESPONSE);
         }
 
         /**
